[sc-12627] Allow to create split strategies

+++ libs/common/src/lib/ai-models/extraction/extraction.component.ts
@@ -1,24 +1,19 @@
 import { ChangeDetectionStrategy, Component, inject, Input } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SisModalService, StickyFooterComponent, TwoColumnsConfigurationItemComponent } from '@nuclia/sistema';
+import { SisModalService, TwoColumnsConfigurationItemComponent } from '@nuclia/sistema';
 import { ModalConfig, PaButtonModule, PaTableModule } from '@guillotinaweb/pastanaga-angular';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { BehaviorSubject, filter, switchMap, take, tap } from 'rxjs';
-import { ExtractConfig, ExtractStrategies, LearningConfigurations } from '@nuclia/core';
+import { ExtractConfig, ExtractStrategies, LearningConfigurations, SplitStrategies, SplitStrategy } from '@nuclia/core';
 import { SDKService } from '@flaps/core';
 import { ExtractionModalComponent } from './extraction-modal/extraction-modal.component';
+import { SplitModalComponent } from './split-modal/split-modal.component';
 
 @Component({
   selector: 'stf-extraction',
-  imports: [
-    CommonModule,
-    PaButtonModule,
-    PaTableModule,
-    StickyFooterComponent,
-    TranslateModule,
-    TwoColumnsConfigurationItemComponent,
-  ],
+  imports: [CommonModule, PaButtonModule, PaTableModule, TranslateModule, TwoColumnsConfigurationItemComponent],
   templateUrl: './extraction.component.html',
+  styleUrls: ['./extraction.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ExtractionComponent {
@@ -27,13 +22,15 @@ export class ExtractionComponent {
   translate = inject(TranslateService);
 
   @Input() learningConfigurations?: LearningConfigurations;
-  strategies = new BehaviorSubject<ExtractStrategies | null>({});
+  extractStrategies = new BehaviorSubject<ExtractStrategies | null>({});
+  splitStrategies = new BehaviorSubject<SplitStrategies>({});
 
   constructor() {
-    this.updateStrategies().subscribe();
+    this.updateExtractStrategies().subscribe();
+    this.updateSplitStrategies().subscribe();
   }
 
-  createStrategy() {
+  createExtractStrategy() {
     this.modalService
       .openModal(
         ExtractionModalComponent,
@@ -47,44 +44,96 @@ export class ExtractionComponent {
             switchMap((kb) => kb.createExtractStrategy(config)),
           ),
         ),
-        switchMap(() => this.updateStrategies()),
+        switchMap(() => this.updateExtractStrategies()),
       )
       .subscribe();
   }
 
-  deleteStrategy(id: string, config: ExtractConfig, event?: Event) {
+  deleteExtractStrategy(id: string, config: ExtractConfig, event?: Event) {
     event?.stopPropagation();
-    return this.modalService
-      .openConfirm({
-        title: 'kb.ai-models.extraction.delete.title',
-        description: this.translate.instant('kb.ai-models.extraction.delete.description', { name: config.name }),
-        confirmLabel: 'generic.delete',
-        isDestructive: true,
-      })
-      .onClose.pipe(
+    return this.confirmDelete(config.name || '')
+      .pipe(
         filter((confirm) => !!confirm),
         switchMap(() => this.sdk.currentKb),
         take(1),
         switchMap((kb) => kb.deleteExtractStrategy(id)),
-        switchMap(() => this.updateStrategies()),
+        switchMap(() => this.updateExtractStrategies()),
       )
       .subscribe();
   }
 
-  updateStrategies() {
+  updateExtractStrategies() {
     return this.sdk.currentKb.pipe(
       take(1),
       switchMap((kb) => kb.getExtractStrategies()),
       tap((strategies) => {
-        this.strategies.next(strategies);
+        this.extractStrategies.next(strategies);
       }),
     );
   }
 
-  displayStrategy(id: string, config: ExtractConfig) {
+  displayExtractStrategy(id: string, config: ExtractConfig) {
     this.modalService.openModal(
       ExtractionModalComponent,
       new ModalConfig({ data: { learningConfigurations: this.learningConfigurations, id, config } }),
     );
+  }
+
+  createSplitStrategy() {
+    this.modalService
+      .openModal(
+        SplitModalComponent,
+        new ModalConfig({ data: { learningConfigurations: this.learningConfigurations } }),
+      )
+      .onClose.pipe(
+        filter((data) => !!data),
+        switchMap((config) =>
+          this.sdk.currentKb.pipe(
+            take(1),
+            switchMap((kb) => kb.createSplitStrategy(config)),
+          ),
+        ),
+        switchMap(() => this.updateSplitStrategies()),
+      )
+      .subscribe();
+  }
+
+  deleteSplitStrategy(id: string, strategy: SplitStrategy, event?: Event) {
+    event?.stopPropagation();
+    return this.confirmDelete(strategy.name || '')
+      .pipe(
+        filter((confirm) => !!confirm),
+        switchMap(() => this.sdk.currentKb),
+        take(1),
+        switchMap((kb) => kb.deleteSplitStrategy(id)),
+        switchMap(() => this.updateSplitStrategies()),
+      )
+      .subscribe();
+  }
+
+  updateSplitStrategies() {
+    return this.sdk.currentKb.pipe(
+      take(1),
+      switchMap((kb) => kb.getSplitStrategies()),
+      tap((strategies) => {
+        this.splitStrategies.next(strategies);
+      }),
+    );
+  }
+
+  displaySplitStrategy(id: string, strategy: SplitStrategy) {
+    this.modalService.openModal(
+      SplitModalComponent,
+      new ModalConfig({ data: { learningConfigurations: this.learningConfigurations, id, config: strategy } }),
+    );
+  }
+
+  private confirmDelete(name: string) {
+    return this.modalService.openConfirm({
+      title: 'kb.ai-models.extraction.delete',
+      description: this.translate.instant('kb.ai-models.extraction.delete-confirm', { name }),
+      confirmLabel: 'generic.delete',
+      isDestructive: true,
+    }).onClose;
   }
 }
